Refactor: Restructure DTO package by moving ResponseObjectJsonDto and ResponseBuilder to common subpackage.
