+++ src/main/java/org/trello/trelloclone/dtos/common/ResponseObjectJsonDto.java
@@ -1,4 +1,4 @@
-package org.trello.trelloclone.dtos;
+package org.trello.trelloclone.dtos.common;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
